+++ kotsu/registration.py
@@ -82,7 +82,7 @@ class _Spec(Generic[Entity]):
     def make(self, **kwargs) -> Entity:
         """Instantiates an instance of the entity."""
         if self.deprecated:
-            raise error.DeprecatedEntityError(
+            raise error.EntityIsDeprecated(
                 f"Attempting to make deprecated entity {self.id}. "
                 "(HINT: is there a newer registered version of this entity?)"
             )