+++ tests/test_registration.py
@@ -74,7 +74,7 @@ def test_make_deprecated_entity():
     registry = registration._Registry()
 
     registry.register("Entity-v0", "fake_entry_point", deprecated=True)
-    with pytest.raises(error.DeprecatedEntityError, match=r"Attempting to make deprecated entity"):
+    with pytest.raises(error.EntityIsDeprecated, match=r"Attempting to make deprecated entity"):
         registry.make("Entity-v0")